clef25/image-retrieval-for-arguments/indexing/requirements.txt
torch  
git+https://github.com/openai/CLIP.git
Pillow
numpy
elasticsearch
pandas